@v0y4g3r 2024-03-11
> Great job! Do you think we could submit these enhancements to the main `prost` project?

No, this optimization is safe if and only if:
- `Buf` is backed by a continuous byte array
- callers can ensure the decode result never outlive original `Bytes`.